refactor: e2e fixures into helper file
